refactor: 즐겨찾기 SwiftData 적용
DetailView 오픈 시 서버에 데이터 전송

사용하지 않는 뷰모델들 삭제
