+++ CleanArea/CleanArea/iOS/Main/View/ListItemView.swift
@@ -17,6 +17,7 @@ struct ListItemView: View {
             VStack(alignment: .leading) {
                 Text(store.policy.polyBizSjnm)
                     .font(.pretendardRegular20)
+                    .multilineTextAlignment(.leading)
                     .foregroundStyle(.black)
                 
                 HStack {
@@ -39,21 +40,18 @@ struct ListItemView: View {
                             Spacer()
                         }
                     }
+                    
+                    Spacer()
+                    
+                    Text(store.category)
+                        .font(.system(size: 15))
+                        .foregroundColor(.gray)
+                        .lineLimit(1)
+                        .minimumScaleFactor(0.5)
+                        .frame(width: UIScreen.main.bounds.width * 0.2)
                 }
             }
-            Spacer()
-            
-            VStack {
-                StarBtn(policy: store.policy)
-                    .padding(2)
-                
-                Text(store.category)
-                    .font(.system(size: 15))
-                    .foregroundColor(.gray)
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.5)
-                    .frame(width: UIScreen.main.bounds.width * 0.2)
-            }
+           
         }
         .padding()
         .background(Color.backgroundGreen)
